Move component to own file
