+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/Summary2/Summary2Component.tsx
@@ -3,7 +3,6 @@ import type { IGenericEditComponent } from '../../componentConfig';
 import { ComponentType } from 'app-shared/types/ComponentType';
 import {
   StudioCard,
-  StudioCombobox,
   StudioNativeSelect,
   StudioParagraph,
   StudioTextfield,
@@ -13,6 +12,7 @@ import { useStudioEnvironmentParams } from 'app-shared/hooks/useStudioEnvironmen
 import { useAppContext } from '../../../../hooks';
 import { useComponentTypeName } from '@altinn/ux-editor/hooks/useComponentTypeName';
 import { useTranslation } from 'react-i18next';
+import { Summmary2ComponentTargetSelector } from './Summary2ComponentTargetSelector';
 
 export const Summary2Component = ({
   schema,
@@ -52,46 +52,48 @@ export const Summary2Component = ({
   };
 
   if (!schema?.properties) return null;
-  const targetSchema = schema.properties.target.properties;
-  const targetProperties = component.target;
+  const target = component.target;
+  const targetTypes = schema.properties.target.properties.type.enum;
+
   return (
     <StudioCard>
       <StudioCard.Header>{t('ux_editor.component_properties.target')}</StudioCard.Header>
-      {/** TODO: Add translation **/}
-      <StudioParagraph size='small'>Target of the summary component</StudioParagraph>
+      <StudioParagraph size='small'>
+        {t('ux_editor.component_properties.target_description')}
+      </StudioParagraph>
       <StudioCard.Content>
         <StudioNativeSelect
-          size='small'
-          label='Type' // TODO: Add translation
-          value={targetProperties.type}
+          size='sm'
+          label={t('ux_editor.component_properties.target_type')}
+          value={target.type}
           onChange={handleTypeChange}
         >
-          {targetSchema.type.enum.map((typeString: string) => (
+          {targetTypes.map((typeString: string) => (
             <option key={typeString} value={typeString}>
               {typeString}
             </option>
           ))}
         </StudioNativeSelect>
-        {targetProperties.type === 'page' && (
-          <Summmary2ComponentTargetId
-            label='page'
-            value={targetProperties.id}
+        {target.type === 'page' && (
+          <Summmary2ComponentTargetSelector
+            label={t('general.page')}
+            value={target.id}
             options={pages}
             onValueChange={handleTargetIdChange}
-          ></Summmary2ComponentTargetId>
+          ></Summmary2ComponentTargetSelector>
         )}
-        {targetProperties.type === 'component' && (
-          <Summmary2ComponentTargetId
-            label='component'
-            value={targetProperties.id}
-            options={filtered}
+        {target.type === 'component' && (
+          <Summmary2ComponentTargetSelector
+            label={t('general.component')}
+            value={target.id}
+            options={components}
             onValueChange={handleTargetIdChange}
-          ></Summmary2ComponentTargetId>
+          ></Summmary2ComponentTargetSelector>
         )}
-        {targetProperties.type === 'layoutSet' && (
+        {target.type === 'layoutSet' && (
           <StudioTextfield
             size='small'
-            label='layoutSet' // TODO: Add translation
+            label={t('general.layout_sets')}
             value={selectedFormLayoutSetName}
             disabled={true}
           />
@@ -101,39 +103,8 @@ export const Summary2Component = ({
   );
 };
 
-type Summary2ComponentTargetIdProps = {
-  label: string;
-  value: string;
-  options: { id: string; description: string }[];
-  onValueChange: (value: string) => void;
-};
-
-const Summmary2ComponentTargetId = ({
-  label,
-  value,
-  options,
-  onValueChange,
-}: Summary2ComponentTargetIdProps) => {
-  const invalidOption = Boolean(value) && !options.some((option) => option.id === value);
-  return (
-    <StudioCombobox
-      size='small'
-      label={label}
-      value={value ? [value] : []}
-      onValueChange={(v) => onValueChange(v[0])}
-      error={invalidOption} // TODO: Add error message
-      multiple={false}
-    >
-      {options.map((option) => (
-        <StudioCombobox.Option value={option.id} key={option.id} description={option.description}>
-          {option.id}
-        </StudioCombobox.Option>
-      ))}
-      {value && invalidOption && (
-        <StudioCombobox.Option disabled value={value} key={value}>
-          {value}
-        </StudioCombobox.Option>
-      )}
-    </StudioCombobox>
-  );
-};
+/**
+ * Added translation keys
+ * ux_editor.component_properties.target_description
+ * ux_editor.component_properties.target_type
+ */
